refactor: replace custom chalk with npm chalk package
